relationship established
HasMany and BelongsTo relationship established between Student and Grade class

+++ example-app/database/migrations/2024_08_17_060544_alter_students_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('students', function (Blueprint $table) {
-            $table-> integer('grade_id')->after('last_name');
+            $table-> unsignedBigInteger('grade_id')->after('last_name');
            
         });
     }
